requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
attrs==23.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.0
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1707444547228/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.8
dlib==19.24.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.5.3
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.3
fire==0.5.0
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-pasta==0.2.0
gradio==4.23.0
gradio_client==0.14.0
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.0
idna==3.6
imageio==2.34.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
importlib_resources==6.4.0
imutils==0.5.4
ipykernel @ file:///D:/bld/ipykernel_1708996677248/work
ipython @ file:///D:/bld/ipython_1709559926914/work
ipywidgets==8.1.2
jax==0.4.25
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1710255804825/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257272359/work
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
lightning-utilities==0.11.1
loguru==0.7.2
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
munch==4.0.0
munkres==1.1.4
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numpy==1.26.1
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
opt-einsum==3.3.0
orjson==3.9.15
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==2.2.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.2.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
plotly==5.20.0
pretrainedmodels==0.7.4
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.20.3
psutil @ file:///D:/bld/psutil_1705722518504/work
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.0
pydub==0.25.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq @ file:///D:/bld/pyzmq_1701783318166/work
referencing==0.34.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.4
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.1
sounddevice==0.4.6
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.36.3
tenacity==8.2.3
threadpoolctl==3.2.0
tifffile==2024.2.12
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==1.3.2
torchvision==0.14.1+cu117
tornado @ file:///D:/bld/tornado_1708363274019/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1710254411456/work
typer==0.10.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
tzdata==2024.1
tzlocal==4.3.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
websockets==11.0.3
widgetsnbextension==4.0.10
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
